Undo FlightScopeTicketHelper
