+++ grpc-api/src/flightTest/java/io/deephaven/grpc_api/flight/FlightMessageRoundTripTest.java
@@ -26,7 +26,7 @@ import io.deephaven.grpc_api.session.SessionState;
 import io.deephaven.grpc_api.session.TicketResolver;
 import io.deephaven.grpc_api.util.FlightExportTicketHelper;
 import io.deephaven.grpc_api.util.Scheduler;
-import io.deephaven.grpc_api.util.FlightScopeTicketHelper;
+import io.deephaven.grpc_api.util.ScopeTicketHelper;
 import io.deephaven.proto.backplane.grpc.HandshakeRequest;
 import io.deephaven.proto.backplane.grpc.HandshakeResponse;
 import io.deephaven.proto.backplane.grpc.SessionServiceGrpc;
@@ -379,7 +379,7 @@ public class FlightMessageRoundTripTest {
     }
 
     private static FlightDescriptor arrowFlightDescriptorForName(String name) {
-        return FlightScopeTicketHelper.descriptorForName(name);
+        return FlightDescriptor.path(ScopeTicketHelper.nameToPath(name));
     }
 
     @Test
